requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.12.2
discord.py==2.3.1
